Finished work on modular props
I modeled modular stone brick wall pieces that work seamlessly when snapping to the grid. I finished up the regular wall, and made a wall with a window. Previous commit had the first wall worked out mostly, but forgot to comment on doing that. Finally I renamed the "TestScene" to "ModularWall" because that is where the wall originated.