Unity edits, Maya tool script, & character UV'd
Large commit for everything from the day. Imported coin to unity, and applied material. Created a MEL script to create UI for user input to duplicate and randomize the position of selected objects(tweaks still need to be made). Finally, UV mapped Mr. Bones character, and applied a basic texture in substand.