Reworked Unity scene & props, as well as retexturing props
This is a large commit of a day's work that I forgot to separate, throughout that time. In terms of the models, the player character(Mr Bones), the weapon rack, and torch have all been re-textured. I then fixed all the textures in Unity, and set up a scene after creating a fire particle system for the torches.